rename the pages
